feat: add labels for stage completion
