+++ src/index.js
@@ -4,6 +4,7 @@ import '../assets/Ancients/index';
 import difficulties from '../data/difficulties';
 import ancientsData from '../data/ancients';
 import { brownCards, blueCards, greenCards } from '../data/mythicCards';
+import { fillDots, stageCards, cardsCount } from './utils'
 
 ///// --- Initialization of Ancient Mode and Dificulty --- //////
 let currentAncient = {};
@@ -21,6 +22,7 @@ let counter = 0;
 const stage1 = document.querySelector('.stage1');
 const stage2 = document.querySelector('.stage2');
 const stage3 = document.querySelector('.stage3');
+const stages = [stage1, stage2, stage3];
 
 const curAncient = document.querySelector('.cur-ancient');
 const diffContainer = document.querySelector('.difficulty-container');
@@ -54,22 +56,32 @@ function resetAll() {
 }
 
 function nextCard() {
-  // console.log(finalCards.pop().cardFace);
-
-  lastCard.style.backgroundImage = `url(${finalCards[finalCards.length-1].cardFace})`;
+  lastCard.style.backgroundImage = `url(${finalCards[finalCards.length - 1].cardFace})`;
   let curCard = {};
   let curStage = {};
   curCard = finalCards.pop();
   counter++;
 
-  lastCardName.innerHTML = curCard.id + curCard.difficulty;
+  lastCardName.innerHTML = curCard.id + ' - ' + curCard.difficulty;
   if (counter <= stageLengths[0]) curStage = stage1
   else if (counter > stageLengths[0] + stageLengths[1]) curStage = stage3
   else curStage = stage2;
   curStage.querySelector('.' + curCard.color).innerHTML--;
-  if (counter == stageLengths[0] + stageLengths[1] + stageLengths[2]) deck.style.visibility = 'hidden';
+  if (counter == stageLengths[0] + stageLengths[1] + stageLengths[2]) {
+    deck.style.visibility = 'hidden';
+    document.querySelector('.shuffle').style.visibility = 'hidden';
+    stage3.style.backgroundColor = 'red';
+    stage3.querySelector('.stage-text').innerHTML += ' completed';
+  } else
+    if (counter == stageLengths[0]) {
+      stage1.style.backgroundColor = 'red';
+      stage1.querySelector('.stage-text').innerHTML += ' completed';
+    } else
+      if (counter == stageLengths[0] + stageLengths[1]) {
+        stage2.style.backgroundColor = 'red';
+        stage2.querySelector('.stage-text').innerHTML += ' completed';
+      }
 }
-
 
 diffContainer.addEventListener('click', diffClick);
 const diffContainerChildren = diffContainer.children;
@@ -80,28 +92,27 @@ for (let i = 0; i <= 4; i++) {
 
 function ancientsClick(e) {
   const tar = e.target;
-//resetAll();
-document.querySelector('.shuffle').style.visibility = 'hidden';
-  
+  resetAll();
+  document.querySelector('.shuffle').style.visibility = 'hidden';
+
   curAncient.innerHTML = tar.id;
   currentAncient = ancientsData.find(el => el.id == tar.id);
-  fillDots(currentAncient);
+  fillDots(currentAncient, stages);
   numberOfCards = cardsCount(currentAncient);
   console.log(numberOfCards);
   document.querySelector('.difficulty-container').style.visibility = 'visible';
 }
 
 function diffClick(e) {
- // resetAll();
- // fillDots(currentAncient);
- // numberOfCards = cardsCount(currentAncient);
+  resetAll();
+  fillDots(currentAncient, stages);
+  numberOfCards = cardsCount(currentAncient);
   let curDiff = e.target;
-  console.log(curDiff);
   diff = e.target.id;
   for (let i = 0; i <= 4; i++) {
     diffContainerChildren[i].classList.remove('active');
   }
-  
+
   curDiff.classList.add('active');
   diffContainer.classList.remove('active');
   console.log(diff);
@@ -112,31 +123,22 @@ function diffClick(e) {
 document.querySelector('.shuffle').addEventListener('click', shuffle);
 
 function shuffle(mode) {
-  // console.log(modifiedSetOfCards(diff));
   deck.style.visibility = 'visible';
   document.querySelector('.deck-container').style.visibility = 'visible';
-  setOfCards = setTotalCards(modifiedSetOfCards(diff), numberOfCards); console.log('setOfCards', setOfCards);
-  finalCards = stageCards(setOfCards, currentAncient);
+
+  stage1.style.backgroundColor = 'white';
+  stage1.querySelector('.stage-text').innerHTML = 'Stage 1';
+  stage2.style.backgroundColor = 'white';
+  stage2.querySelector('.stage-text').innerHTML = 'Stage 2';
+  stage3.style.backgroundColor = 'white';
+  stage3.querySelector('.stage-text').innerHTML = 'Stage 3';
+
+  setOfCards = setTotalCards(modifiedSetOfCards(diff), numberOfCards);
+  finalCards = stageCards(setOfCards, currentAncient, stageLengths);
   console.log('finalCards', finalCards);
+  lastCard.style.backgroundImage = 'none';
+  lastCardName.innerHTML = '';
 }
-
-function totalCards(ancient) {
-  return (
-    ancient.firstStage.greenCards + ancient.firstStage.blueCards + ancient.firstStage.brownCards +
-    ancient.secondStage.greenCards + ancient.secondStage.blueCards + ancient.secondStage.brownCards +
-    ancient.thirdStage.greenCards + ancient.thirdStage.blueCards + ancient.thirdStage.brownCards
-  )
-}
-
-///// --- CARDS COUNT OF EVERY COLOR --- /////
-function cardsCount(ancient) {
-  return {
-    greenCards: ancient.firstStage.greenCards + ancient.secondStage.greenCards + ancient.thirdStage.greenCards,
-    blueCards: ancient.firstStage.blueCards + ancient.secondStage.blueCards + ancient.thirdStage.blueCards,
-    brownCards: ancient.firstStage.brownCards + ancient.secondStage.brownCards + ancient.thirdStage.brownCards,
-  }
-}
-//////////////////////////////////////////////////////////////////////////////
 
 function modifiedSetOfCards(diff) {
   switch (diff) {
@@ -242,96 +244,3 @@ function setTotalCards(cards, colorsCount) {
     blueCards: blue
   }
 }
-
-function stageCards(cards, curAncient) {
-  const finalCards = {
-    stage1: {},
-    stage2: {},
-    stage3: {},
-    cards: []
-  };
-  const curCards = {};
-  Object.assign(curCards, cards);
-  //console.log(curAncient);
-
-  /////////// FIRST STAGE //////////////
-  curCards.greenCards = _.shuffle(curCards.greenCards);
-  finalCards.stage1.greenCards = curCards.greenCards.slice(0, curAncient.firstStage.greenCards);
-  curCards.greenCards = curCards.greenCards.slice(curAncient.firstStage.greenCards, curCards.greenCards.length);
-
-  curCards.brownCards = _.shuffle(curCards.brownCards);
-  finalCards.stage1.brownCards = curCards.brownCards.slice(0, curAncient.firstStage.brownCards);
-  curCards.brownCards = curCards.brownCards.slice(curAncient.firstStage.brownCards, curCards.brownCards.length);
-
-  curCards.blueCards = _.shuffle(curCards.blueCards);
-  finalCards.stage1.blueCards = curCards.blueCards.slice(0, curAncient.firstStage.blueCards);
-  curCards.blueCards = curCards.blueCards.slice(curAncient.firstStage.blueCards, curCards.blueCards.length);
-
-  finalCards.stage1.cards = finalCards.stage1.greenCards.slice();
-  finalCards.stage1.cards.push(...finalCards.stage1.brownCards);
-  finalCards.stage1.cards.push(...finalCards.stage1.blueCards);
-  finalCards.stage1.cards = _.shuffle(finalCards.stage1.cards);
-  stageLengths[0] = finalCards.stage1.cards.length;
-
-  /////////// SECOND STAGE //////////////
-  curCards.greenCards = _.shuffle(curCards.greenCards);
-  finalCards.stage2.greenCards = curCards.greenCards.slice(0, curAncient.secondStage.greenCards);
-  curCards.greenCards = curCards.greenCards.slice(curAncient.secondStage.greenCards, curCards.greenCards.length);
-
-  curCards.brownCards = _.shuffle(curCards.brownCards);
-  finalCards.stage2.brownCards = curCards.brownCards.slice(0, curAncient.secondStage.brownCards);
-  curCards.brownCards = curCards.brownCards.slice(curAncient.secondStage.brownCards, curCards.brownCards.length);
-
-  curCards.blueCards = _.shuffle(curCards.blueCards);
-  finalCards.stage2.blueCards = curCards.blueCards.slice(0, curAncient.secondStage.blueCards);
-  curCards.blueCards = curCards.blueCards.slice(curAncient.secondStage.blueCards, curCards.blueCards.length);
-
-  finalCards.stage2.cards = finalCards.stage2.greenCards.slice();
-  finalCards.stage2.cards.push(...finalCards.stage2.brownCards);
-  finalCards.stage2.cards.push(...finalCards.stage2.blueCards);
-  finalCards.stage2.cards = _.shuffle(finalCards.stage2.cards);
-  stageLengths[1] = finalCards.stage2.cards.length;
-
-  /////////// THIRD STAGE //////////////
-  curCards.greenCards = _.shuffle(curCards.greenCards);
-  finalCards.stage3.greenCards = curCards.greenCards.slice(0, curAncient.thirdStage.greenCards);
-  curCards.greenCards = curCards.greenCards.slice(curAncient.thirdStage.greenCards, curCards.greenCards.length);
-
-  curCards.brownCards = _.shuffle(curCards.brownCards);
-  finalCards.stage3.brownCards = curCards.brownCards.slice(0, curAncient.thirdStage.brownCards);
-  curCards.brownCards = curCards.brownCards.slice(curAncient.thirdStage.brownCards, curCards.brownCards.length);
-
-  curCards.blueCards = _.shuffle(curCards.blueCards);
-  finalCards.stage3.blueCards = curCards.blueCards.slice(0, curAncient.thirdStage.blueCards);
-  curCards.blueCards = curCards.blueCards.slice(curAncient.thirdStage.blueCards, curCards.blueCards.length);
-
-  finalCards.stage3.cards = finalCards.stage3.greenCards.slice();
-  finalCards.stage3.cards.push(...finalCards.stage3.brownCards);
-  finalCards.stage3.cards.push(...finalCards.stage3.blueCards);
-  finalCards.stage3.cards = _.shuffle(finalCards.stage3.cards);
-  stageLengths[2] = finalCards.stage3.cards.length;
-
-  finalCards.cards = [...finalCards.stage3.cards];
-  finalCards.cards.push(...finalCards.stage2.cards);
-  finalCards.cards.push(...finalCards.stage1.cards);
-  console.log(finalCards);
-  return finalCards.cards;
-}
-
-
-////////// --- FILL DOTS OF EVERY STAGE --- //////////
-function fillDots(ancient) {
-
-  stage1.querySelector('.green').innerHTML = ancient.firstStage.greenCards;
-  stage1.querySelector('.blue').innerHTML = ancient.firstStage.blueCards;
-  stage1.querySelector('.brown').innerHTML = ancient.firstStage.brownCards;
-
-  stage2.querySelector('.green').innerHTML = ancient.secondStage.greenCards;
-  stage2.querySelector('.blue').innerHTML = ancient.secondStage.blueCards;
-  stage2.querySelector('.brown').innerHTML = ancient.secondStage.brownCards;
-
-  stage3.querySelector('.green').innerHTML = ancient.thirdStage.greenCards;
-  stage3.querySelector('.blue').innerHTML = ancient.thirdStage.blueCards;
-  stage3.querySelector('.brown').innerHTML = ancient.thirdStage.brownCards;
-}
-//////////////////////////////////////////////////////////////////////////////
